requirements.txt
paho-mqtt
gpiod
